refactor: extract reusable task composer
